Move Pi tests into dedicated bignum/tests/pi_test.cpp
Extract the Pi-generation tests (Machin's formula and Gauss-Legendre)
and their helper functions out of tests.cpp into a new piTest target.
The sqrt building-block tests stay in tests.cpp. Also add a test that
the two Pi algorithms agree to 100 digits.

No behavior change; pure test reorganization.
